Added unittests for IntegratedActionModelLPF + fix some warnings

+++ tests/factory/statelpf.cpp
@@ -34,7 +34,7 @@ std::vector<std::string> LPFJointListFactory::create_names(
     case LPFJointMaskType::ALL: {
       for (std::vector<std::string>::iterator iter = model->names.begin();
            iter != model->names.end(); ++iter) {
-        if (model->getJointId(*iter) < model->njoints &&
+        if (static_cast<int>(model->getJointId(*iter)) < model->njoints &&
             model->nvs[model->getJointId(*iter)] == 1) {
           lpf_joint_names.push_back(*iter);
         }
@@ -62,9 +62,9 @@ std::vector<int> LPFJointListFactory::create_ids(
     case LPFJointMaskType::ALL: {
       for (std::vector<std::string>::iterator iter = model->names.begin();
            iter != model->names.end(); ++iter) {
-        if (model->getJointId(*iter) < model->njoints &&
+        if (static_cast<int>(model->getJointId(*iter)) < model->njoints &&
             model->nvs[model->getJointId(*iter)] == 1) {
-          lpf_joint_ids.push_back(model->getJointId(*iter));
+          lpf_joint_ids.push_back(static_cast<int>(model->getJointId(*iter)));
         }
       }
       break;
